refactor : Transfer root endpoint to a test folder and controller
