+++ backend/src/main/java/ch/sectioninformatique/template/item/ItemController.java
@@ -1,7 +1,6 @@
 package ch.sectioninformatique.template.item;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,25 +21,6 @@ public class ItemController {
 
     @Autowired
     private ItemService itemService;
-
-    @Autowired
-    private Environment environment;
-
-    /**
-     * Returns system information and environment variables.
-     * This endpoint is used to verify that the application is running
-     * and to display configuration information.
-     *
-     * @return A formatted string containing system information
-     */
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/")
-    public String getHello() {
-        return "<strong>Hello World !</strong><br>" +
-               "<strong>JAVA_HOME : </strong>" + environment.getProperty("JAVA_HOME") + "<br>" +
-               "<strong>Spring active profile : </strong>" + environment.getProperty("spring.profiles.active") + "<br>" +
-               "<strong>Database used : </strong>" + environment.getProperty("spring.datasource.url");
-    }
 
     /**
      * Retrieves all items in the system.
